requirements.txt
albumentations==1.2.0
hydra-core==1.2.0
jittor==1.3.4.16
numpy==1.23.0
omegaconf==2.2.2
opencv_python_headless==4.6.0.66
Pillow==9.2.0
pyrsistent==0.18.1
tqdm==4.64.0
wandb==0.12.20
